+++ polymer-modulizer/src/cli/command-workspace.ts
@@ -19,9 +19,9 @@ import {Workspace} from 'polymer-workspaces';
 
 import {CliOptions} from '../cli';
 import convertWorkspace from '../convert-workspace';
-import { testWorkspace, testWorkspaceInstallOnly } from '../test-workspace';
-import githubPushWorkspace from '../push-workspace';
 import npmPublishWorkspace from '../publish-workspace';
+import githubPushWorkspace from '../push-workspace';
+import {testWorkspace, testWorkspaceInstallOnly} from '../test-workspace';
 import {logStep} from '../util';
 
 const githubTokenMessage = `
@@ -129,6 +129,7 @@ export default async function run(options: CliOptions) {
     npmImportStyle: options['import-style'],
     packageVersion: npmPackageVersion,
     reposToConvert,
+    deleteFiles: options['delete-files'],
   });
 
   logStep(3, 3, '🎉', `Conversion Complete!`);