+++ polymer-modulizer/src/tools/update-fixtures.ts
@@ -17,19 +17,17 @@ require('source-map-support').install();
 
 import * as path from 'path';
 import {exec} from 'mz/child_process';
-import * as fs from 'mz/fs';
-import * as rimraf from 'rimraf';
-import util = require('util');
-import _mkdirp = require('mkdirp');
-const mkdirp = util.promisify(_mkdirp);
+import * as fs from 'fs-extra';
 
 import convertPackage from '../convert-package';
 
 interface UpdateFixtureOptions {
   folder: string;
   repoUrl: string;
+  branch?: string;
   packageName: string;
   packageVersion: string;
+  deleteFiles?: string[];
 }
 
 async function updateFixture(options: UpdateFixtureOptions) {
@@ -37,30 +35,36 @@ async function updateFixture(options: UpdateFixtureOptions) {
       path.resolve(__dirname, '../../fixtures/packages/', options.folder);
   const sourceDir = path.join(fixturesDir, 'source');
   const convertedDir = path.join(fixturesDir, 'expected');
+  const branch = options.branch || 'master';
 
-  console.log(`Cloning ${options.repoUrl} to ${sourceDir}...`);
-  await mkdirp(fixturesDir);
-  rimraf.sync(sourceDir);
+  console.log(`Cloning ${options.repoUrl} #${branch} to ${sourceDir}...`);
+  await fs.ensureDir(fixturesDir);
+  await fs.remove(sourceDir);
 
   await exec(
-      `git clone ${options.repoUrl} ${sourceDir} --depth=1`,
+      `git clone ${options.repoUrl} ${sourceDir} --branch=${branch} --depth=1`,
       {cwd: fixturesDir});
-  rimraf.sync(path.join(sourceDir, '.git'));
-  rimraf.sync(path.join(sourceDir, '.github'));
-  rimraf.sync(path.join(sourceDir, '.gitignore'));
+  await fs.remove(path.join(sourceDir, '.git'));
+  await fs.remove(path.join(sourceDir, '.github'));
+  await fs.remove(path.join(sourceDir, '.gitignore'));
 
   await overridePolymer(sourceDir);
 
   await exec('bower install', {cwd: sourceDir});
 
+  // We're going to do an in-place conversion.
+  await fs.emptyDir(convertedDir);
+  await fs.copy(sourceDir, convertedDir);
+
   console.log(`Converting...`);
   await convertPackage({
-    inDir: sourceDir,
+    inDir: convertedDir,
     outDir: convertedDir,
-    cleanOutDir: true,
+    cleanOutDir: false,
     packageName: options.packageName,
     packageVersion: options.packageVersion,
     addImportPath: true,
+    deleteFiles: options.deleteFiles,
   });
   console.log(`Done.`);
 }
@@ -85,12 +89,15 @@ async function overridePolymer(sourceDir: string) {
 }
 
 (async () => {
+  let exitCode = 0;
+
   await Promise.all([
     updateFixture({
       folder: 'polymer',
       repoUrl: 'https://github.com/Polymer/polymer.git',
       packageName: '@polymer/polymer',
       packageVersion: '3.0.0',
+      deleteFiles: ['types'],
     }),
     updateFixture({
       folder: 'paper-button',
@@ -104,5 +111,11 @@ async function overridePolymer(sourceDir: string) {
       packageName: '@polymer/iron-icon',
       packageVersion: '3.0.0',
     }),
-  ]);
+  ].map((p) => p.catch((e) => {
+    // Exit with an error code if any fixture fails, but let them all finish.
+    console.error(e);
+    exitCode = 1;
+  })));
+
+  process.exit(exitCode);
 })();